fix: eliminate 364 build warnings (512→148) + wire geometry into embodied sim & LNN gradient
Warning fixes (71% reduction, 512→148 unique warnings):
- Implicit function declarations: 77→1 (added stdio.h to 25 files, forward
  declarations for set_error/brain_immune_*/bbb_validate_data/gw_cognitive_*/
  core_directives_*/lnn_network_backward, fixed typo nimcp_platform_time→nimcp_time)
- Builtin declaration mismatches: 107→0 (stdlib.h/string.h/math.h includes)
- Macro redefinitions: 10→0 (#ifndef guards for NIMCP_PATH_BUFFER_SIZE,
  NIMCP_MAX_LAYERS, NIMCP_CHECK_NULL, NIMCP_SAFE_STRCPY, time constants;
  #undef for BRIDGE_LOCK/UNLOCK, LOG_MODULE)
- Int conversion: 49→2 (return NULL→-1 in int functions, return -1→NULL
  in pointer functions, fixed lgss_loader 24 return NULLs)
- Incompatible pointer types: 164→51 (STDP bridge mutex API migration
  platform→thread layer, heartbeat_instance void* first arg, brain_get_brain_regions
  return type brain_module_t*, visual_cortex_process uint8_t* cast,
  consolidation handle type, fuzzy kg_wiring cast, parietal sleep &)
- Pointer-to-int cast: 3→0 (uintptr_t double-cast in logging macros)
- kg_gc_agent struct forward declaration for callback typedefs
- metabolic tensor create: NIMCP_API_CHECK→NIMCP_API_CHECK_RET_NULL

Geometry integration (3 new modules wired):
- Embodied simulation: SO(3) geodesic distance for rotation metabolic cost,
  SO(3) SLERP orientation interpolation during ROTATE action execution
- LNN gradient: Riemannian parallel transport for adjoint step when dim≤64,
  Fisher metric G=J^TJ from Jacobian, Christoffel symbols for curvature-
  aware gradient transport (falls back to flat Euler for larger dimensions)

+++ include/utils/logging/nimcp_logging.h
@@ -1003,23 +1003,23 @@ NIMCP_EXPORT bool nimcp_log_is_tty(void);
  */
 #define nimcp_log_debug(module_id, fmt, ...) \
     nimcp_log_write(NULL, LOG_LEVEL_DEBUG, NULL, __FILE__, __LINE__, \
-                    "[0x%04X] " fmt, (unsigned int)(module_id), ##__VA_ARGS__)
+                    "[0x%04X] " fmt, (unsigned int)(uintptr_t)(module_id), ##__VA_ARGS__)
 
 #define nimcp_log_info(module_id, fmt, ...) \
     nimcp_log_write(NULL, LOG_LEVEL_INFO, NULL, __FILE__, __LINE__, \
-                    "[0x%04X] " fmt, (unsigned int)(module_id), ##__VA_ARGS__)
+                    "[0x%04X] " fmt, (unsigned int)(uintptr_t)(module_id), ##__VA_ARGS__)
 
 #define nimcp_log_warning(module_id, fmt, ...) \
     nimcp_log_write(NULL, LOG_LEVEL_WARN, NULL, __FILE__, __LINE__, \
-                    "[0x%04X] " fmt, (unsigned int)(module_id), ##__VA_ARGS__)
+                    "[0x%04X] " fmt, (unsigned int)(uintptr_t)(module_id), ##__VA_ARGS__)
 
 #define nimcp_log_warn(module_id, fmt, ...) \
     nimcp_log_write(NULL, LOG_LEVEL_WARN, NULL, __FILE__, __LINE__, \
-                    "[0x%04X] " fmt, (unsigned int)(module_id), ##__VA_ARGS__)
+                    "[0x%04X] " fmt, (unsigned int)(uintptr_t)(module_id), ##__VA_ARGS__)
 
 #define nimcp_log_error(module_id, fmt, ...) \
     nimcp_log_write(NULL, LOG_LEVEL_ERROR, NULL, __FILE__, __LINE__, \
-                    "[0x%04X] " fmt, (unsigned int)(module_id), ##__VA_ARGS__)
+                    "[0x%04X] " fmt, (unsigned int)(uintptr_t)(module_id), ##__VA_ARGS__)
 
 /**
  * @brief Get a named logger (alias for nimcp_log_get_global for compatibility)
